# backend/docs/frontend/openapi.yaml
openapi: 3.0.3
info:
  title: MedRack Operator API
  description: |
    Stable JSON API for MedRack's backend services. Exposes 22 endpoints
    under `/api/v1/*` for frontend integration.

    This is the v1 contract, frozen as of v0.3.0-backend-freeze.
    Future versions (v2) will be added in parallel; v1 will not change.

    See `/docs/frontend/API_REFERENCE.md` for the full reference.
  version: 1.0.0
  contact:
    name: MedRack Backend
    url: https://github.com/medrack
  license:
    name: Personal Edition
servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://localhost:8000/api/v1
    description: Local development (with API prefix)
...
paths:
...
  /api/v1/library/books:
    get:
      tags: [library]
      summary: List all textbooks
      operationId: listBooks
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookInfo'
    post:
      tags: [library]
      summary: Add a book by ingesting a PDF
      operationId: addBook
      parameters:
        - name: pdf_path
          in: query
          required: true
          schema:
            type: string
        - name: subject
          in: query
          required: true
          schema:
            type: string
            enum: [psm, fmt]
        - name: book_title
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddBookResponse'
  /api/v1/library/books/{book_id}:
    delete:
      tags: [library]
      summary: Remove a book (soft delete)
      operationId: removeBook
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveBookResponse'
  /api/v1/library/books/{book_id}/reindex:
    post:
      tags: [library]
      summary: Re-index a book
      operationId: reindexBook
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddBookResponse'
  /api/v1/library/ingestion-status/{book_id}:
    get:
      tags: [library]
      summary: Get ingestion status for a book
      operationId: getIngestionStatus
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionStatus'
...
  /api/v1/questions/generate:
    post:
      tags: [questions]
      summary: Generate a single answer
      operationId: generateAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationResult'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/questions/batch:
    post:
      tags: [questions]
      summary: Generate answers for a batch of questions
      operationId: generateBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchGenerateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenerationResult'
  /api/v1/questions/{qid}/revise:
    post:
      tags: [questions]
      summary: Revise an existing answer
      operationId: reviseAnswer
      parameters:
        - name: qid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviseRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationResult'
  /api/v1/questions/stale:
    get:
      tags: [questions]
      summary: List (or re-answer) stale cached answers
      operationId: listStaleAnswers
      parameters:
        - name: module_name
          in: query
          required: true
          schema:
            type: string
        - name: dry_run
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaleResponse'
  /api/v1/pipeline/inspect:
    get:
      tags: [pipeline]
      summary: Inspect all 6 pipeline stages for a question
      operationId: inspectPipeline
      parameters:
        - name: qid
          in: query
          required: true
          schema:
            type: string
        - name: question_text
          in: query
          required: true
          schema:
            type: string
        - name: subject
          in: query
          required: true
          schema:
            type: string
            enum: [psm, fmt]
        - name: marks
          in: query
          required: false
          schema:
            type: integer
            default: 10
            enum: [5, 10, 15]
        - name: question_type
          in: query
          required: false
          schema:
            type: string
            default: theory
            enum: [theory, mcq]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineTrace'
  /api/v1/validation/validate:
    post:
      tags: [validation]
      summary: Run the validation pipeline against an answer
      operationId: validateAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationReport'
...
  /api/v1/benchmarks/runs/{run_id}:
    get:
      tags: [benchmarks]
      summary: Get a full benchmark run report
      operationId: getBenchmarkRun
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/RunNotFound'
  /api/v1/benchmarks/compare:
    get:
      tags: [benchmarks]
      summary: Compare two benchmark runs
      operationId: compareBenchmarkRuns
      parameters:
        - name: run_a
          in: query
          required: true
          schema:
            type: string
        - name: run_b
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkCompareResponse'
  /api/v1/cache/entries:
    get:
      tags: [cache]
      summary: List cache entries
      operationId: listCacheEntries
      parameters:
        - name: subject
          in: query
          required: false
          schema:
            type: string
            enum: [psm, fmt]
        - name: stale_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CacheEntry'
  /api/v1/cache/entries/{qid}:
    get:
      tags: [cache]
      summary: Get a single cache entry by qid
      operationId: getCacheEntry
      parameters:
        - name: qid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/CacheEntryNotFound'
...
  /api/v1/cache/reanswer:
    post:
      tags: [cache]
      summary: Mark a cache entry as stale
      operationId: reanswerCache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReanswerRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReanswerResponse'
  /api/v1/logs/{name}:
    get:
      tags: [logs]
      summary: Tail a log file
      operationId: tailLog
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            enum: [ingestion, generation, validation, benchmark]
        - name: n
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '400':
          $ref: '#/components/responses/UnknownLog'
  /api/v1/logs/{name}/search:
    get:
      tags: [logs]
      summary: Search a log file
      operationId: searchLog
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            enum: [ingestion, generation, validation, benchmark]
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: n
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '400':
          $ref: '#/components/responses/UnknownLog'
components:
  responses:
...
    ValidationError:
      description: Validation error (FastAPI default)
      content:
        application/json:
          schema:
            type: object
  schemas:
...
    BookInfo:
      type: object
      required: [schema_version, book_id, title, subject, path, indexed]
      properties:
        schema_version:
          type: integer
        book_id:
          type: string
        title:
          type: string
        subject:
          type: string
          enum: [psm, fmt, unknown]
        path:
          type: string
        indexed:
          type: boolean
        indexed_at:
          type: string
          nullable: true
        chunk_count:
          type: integer
    AddBookResponse:
      type: object
      required: [ok, book_id, path, subject]
      properties:
        ok:
          type: boolean
        book_id:
          type: string
        path:
          type: string
        subject:
          type: string
    RemoveBookResponse:
      type: object
      required: [ok, book_id, moved_to]
      properties:
        ok:
          type: boolean
        book_id:
          type: string
        moved_to:
          type: string
...
    IngestionStatus:
      type: object
      required: [schema_version, book_id, status, started_at]
      properties:
        schema_version:
          type: integer
        book_id:
          type: string
        status:
          type: string
          enum: [unknown, pending, running, succeeded, failed]
        started_at:
          type: string
        finished_at:
          type: string
          nullable: true
        chunk_count:
          type: integer
        error:
          type: string
          nullable: true
    GenerateRequest:
      type: object
      required: [qid, question_text, subject, marks]
      properties:
        qid:
          type: string
        question_text:
          type: string
        subject:
          type: string
          enum: [psm, fmt]
        marks:
          type: integer
          default: 10
          enum: [5, 10, 15]
        question_type:
          type: string
          default: theory
          enum: [theory, mcq]
        book_id:
          type: string
        chapter:
          type: string
    BatchGenerateRequest:
      type: object
      required: [requests]
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/GenerateRequest'
    ReviseRequest:
      type: object
      required: [subject, revised_question_text]
      properties:
        subject:
          type: string
          enum: [psm, fmt]
        revised_question_text:
          type: string
    GenerationResult:
      type: object
      required: [schema_version, qid, ok, cache_hit, token_count, latency_seconds]
      properties:
        schema_version:
          type: integer
        qid:
          type: string
        ok:
          type: boolean
        answer_text:
          type: string
          nullable: true
        pdf_path:
          type: string
          nullable: true
        cache_hit:
          type: boolean
        error:
          type: string
          nullable: true
        token_count:
          type: integer
        latency_seconds:
          type: number
    StaleResponse:
      type: object
      required: [ok, dry_run]
      properties:
        ok:
          type: boolean
        dry_run:
          type: boolean
        stale_count:
          type: integer
        stale_qids:
          type: array
          items:
            type: string
        reanswered_count:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/GenerationResult'
...
    ReanswerRequest:
      type: object
      required: [qid]
      properties:
        qid:
          type: string
...